+++ packages/react-pointcuts/src/withTogglePointFactory/getComponent/index.js
@@ -1,52 +1,49 @@
 import withPlugins from "./withPlugins";
 import withErrorBoundary from "./withErrorBoundary";
 import { forwardRef } from "react";
+import getDisplayName from "../getDisplayName";
 
 const getControlOrVariant = ({
   matchedFeatures,
   matchedVariant,
   logError,
-  control
+  packedBaseModule,
+  unpackComponent
 }) => {
-  if (!matchedFeatures.length) {
-    return control;
+  if (!matchedFeatures.length || !matchedVariant) {
+    return unpackComponent(packedBaseModule);
   }
 
-  let Component = control;
-  if (matchedVariant) {
-    const { codeRequest, variables } = matchedVariant;
-    const { default: VariantWithoutVariables } = codeRequest;
-    const Variant = forwardRef((props, ref) => (
-      <VariantWithoutVariables {...{ ...variables, ...props, ref }} />
-    ));
-    Variant.displayName = `Variant(${
-      VariantWithoutVariables.displayName ||
-      VariantWithoutVariables.name ||
-      "Component"
-    })`;
+  const { packedModule, variables } = matchedVariant;
+  const VariantWithoutVariables = unpackComponent(packedModule);
+  const Variant = forwardRef((props, ref) => (
+    <VariantWithoutVariables {...{ ...variables, ...props, ref }} />
+  ));
+  Variant.displayName = `Variant(${getDisplayName(VariantWithoutVariables)})`;
 
-    Component = withErrorBoundary({
-      Variant,
-      logError,
-      fallback: control
-    });
-  }
-  return Component;
+  const component = withErrorBoundary({
+    Variant,
+    logError,
+    packedBaseModule,
+    unpackComponent
+  });
+
+  return component;
 };
 
 const getComponent = (params) => {
-  let Component = getControlOrVariant(params);
+  let component = getControlOrVariant(params);
 
   const { plugins, ...rest } = params;
   if (plugins) {
-    Component = withPlugins({
-      Component,
+    component = withPlugins({
+      component,
       plugins,
       ...rest
     });
   }
 
-  return Component;
+  return component;
 };
 
 export default getComponent;