+++ packages/react-pointcuts/src/withTogglePointFactory/getComponent/index.test.js
@@ -3,14 +3,23 @@ import withErrorBoundary from "./withErrorBoundary";
 import withPlugins from "./withPlugins";
 import { render, screen } from "@testing-library/react";
 import { createRef, forwardRef } from "react";
+import getDisplayName from "../getDisplayName";
 
 jest.mock("./withErrorBoundary", () => jest.fn());
 jest.mock("./withPlugins", () => jest.fn());
+const mockDisplayName = "test-display-name";
+jest.mock("../getDisplayName", () => jest.fn(() => mockDisplayName));
 
 const mockVariantComponent = "test-variant-component";
 const MockVariantComponent = forwardRef(
   jest.fn((_, ref) => <div ref={ref} data-testid={mockVariantComponent} />)
 );
+
+const unpackMarker = Symbol("test-unpack-marker");
+const unpackComponent = jest.fn().mockImplementation((module) => {
+  module[unpackMarker] = true;
+  return module;
+});
 
 describe("getComponent", () => {
   let result, params;
@@ -20,12 +29,13 @@ describe("getComponent", () => {
     jest.clearAllMocks();
     params = {
       logError: Symbol("test-error-logger"),
-      control: () => Symbol("test-base-component"),
-      plugins: Symbol("test-plugins")
+      packedBaseModule: () => Symbol("test-base-component"),
+      plugins: Symbol("test-plugins"),
+      unpackComponent
     };
-    withPlugins.mockImplementation(({ Component }) => {
-      Component[pluginMarker] = true;
-      return Component;
+    withPlugins.mockImplementation(({ component }) => {
+      component[pluginMarker] = true;
+      return component;
     });
   });
 
@@ -33,7 +43,7 @@ describe("getComponent", () => {
     it("should run plugins on the the matched features, passing the params that were passed to the component, in-case the plugins need them", () => {
       const { plugins, ...rest } = params;
       expect(withPlugins).toHaveBeenCalledWith({
-        Component: result,
+        component: result,
         plugins,
         ...rest
       });
@@ -42,8 +52,10 @@ describe("getComponent", () => {
   };
 
   const makeFallbackAssertion = () => {
-    it("should return the control (base) component", () => {
-      expect(result).toBe(params.control);
+    it("should return the unpacked base component", () => {
+      expect(unpackComponent).toHaveBeenCalledWith(params.packedBaseModule);
+      expect(result).toBe(params.packedBaseModule);
+      expect(result[unpackMarker]).toBe(true);
     });
   };
 
@@ -100,7 +112,7 @@ describe("getComponent", () => {
         params = {
           ...params,
           matchedVariant: {
-            codeRequest: { default: MockVariantComponent },
+            packedModule: MockVariantComponent,
             variables
           }
         };
@@ -108,18 +120,25 @@ describe("getComponent", () => {
       });
 
       it("should wrap the variant with an error boundary, to ensure errors in the variant result in falling back to the base/default component", () => {
-        const { control: Component } = params;
+        const { packedBaseModule: Component } = params;
         expect(withErrorBoundary).toHaveBeenCalledWith({
           logError: params.logError,
           Variant: expect.anything(),
-          fallback: Component
+          packedBaseModule: Component,
+          unpackComponent
         });
         expect(result[errorBoundariedMarker]).toBe(true);
       });
 
+      it("should set a display name on the variant", () => {
+        const [[{ Variant }]] = withErrorBoundary.mock.calls;
+        expect(getDisplayName).toHaveBeenCalledWith(MockVariantComponent);
+        expect(Variant.displayName).toEqual(`Variant(${mockDisplayName})`);
+      });
+
       makeCommonAssertions();
 
-      describe("when the returned variant is rendered", () => {
+      describe("when the returned component is rendered", () => {
         const componentProps = {
           "test-component-prop": Symbol("test-component-prop-value"),
           [innateProp]: Symbol("test-innate-prop-value")
@@ -131,7 +150,14 @@ describe("getComponent", () => {
           render(<Component {...componentProps} ref={mockRef} />);
         });
 
-        it("should return the variant, sending down any innate props, attaching the passed ref, and any variables from the feature as props (preferring innate props over variables - so features can't overwrite innate/in-built props)", () => {
+        it("should return the unpacked variant component", () => {
+          expect(unpackComponent).toHaveBeenCalledWith(
+            params.matchedVariant.packedModule
+          );
+          expect(MockVariantComponent[unpackMarker]).toBe(true);
+        });
+
+        it("should render the variant, sending down any innate props, attaching the passed ref, and any variables from the feature as props (preferring innate props over variables - so features can't overwrite innate/in-built props)", () => {
           const variantElement = screen.getByTestId(mockVariantComponent);
           expect(variantElement).toBeInTheDocument();
           expect(variantElement).toBe(mockRef.current);