+++ packages/react-pointcuts/src/withTogglePointFactory/useDeferredValueWhereAvailable.js
@@ -1,0 +1,10 @@
+let useDeferredValue = (value) => value;
+
+(async () => {
+  const React = await import("react");
+  if (React.useDeferredValue) {
+    useDeferredValue = React.useDeferredValue;
+  }
+})();
+
+export { useDeferredValue };