+++ packages/react-pointcuts/src/withToggledHookFactory/index.test.js
@@ -20,16 +20,23 @@ describe("withToggledHookFactory", () => {
   const mockPlugins = [Symbol("test-plugin1"), Symbol("test-plugin2")];
   const initialProps = Symbol("test-arg");
   const mockActiveFeatures = Symbol("test-active-features");
+  const variantKey = Symbol("test-variant-key");
   const getActiveFeatures = jest.fn(() => mockActiveFeatures);
+  const unpack = jest.fn((module) => module);
 
   beforeEach(() => {
     jest.clearAllMocks();
     jest.mock("../useCodeMatches", () => jest.fn(() => mockMatches));
     const withToggledHook = withToggledHookFactory({
       getActiveFeatures,
+      variantKey,
       plugins: mockPlugins
     });
-    toggledHook = withToggledHook({ default: mockControlHook }, featuresMap);
+    toggledHook = withToggledHook({
+      joinPoint: { default: mockControlHook },
+      featuresMap,
+      unpack
+    });
   });
 
   it("should get code selection plugins", () => {
@@ -49,6 +56,7 @@ describe("withToggledHookFactory", () => {
       it("should get code matches", () => {
         expect(useCodeMatches).toHaveBeenCalledWith({
           featuresMap,
+          variantKey,
           activeFeatures: mockActiveFeatures
         });
       });
@@ -62,13 +70,19 @@ describe("withToggledHookFactory", () => {
 
       beforeEach(() => {
         mockMatches.matchedVariant = {
-          codeRequest: {
+          packedModule: {
             default: variant
           }
         };
         ({ result } = renderHook(toggledHook, {
           initialProps
         }));
+      });
+
+      it("should unpack the variant module", () => {
+        expect(unpack).toHaveBeenCalledWith(
+          mockMatches.matchedVariant.packedModule
+        );
       });
 
       it("should call and return the output of the matched variant", () => {
@@ -92,6 +106,12 @@ describe("withToggledHookFactory", () => {
         ({ result } = renderHook(toggledHook, { initialProps }));
       });
 
+      it("should unpack the control module", () => {
+        expect(unpack).toHaveBeenCalledWith({
+          default: mockControlHook
+        });
+      });
+
       it("should call and return the output of the fallback (control) hook", () => {
         expect(mockControlHook).toHaveBeenCalledWith(initialProps);
         expect(result.current).toBe(output);
@@ -106,7 +126,7 @@ describe("withToggledHookFactory", () => {
 
     beforeEach(() => {
       mockMatches.matchedVariant = {
-        codeRequest: {
+        packedModule: {
           default: jest.fn()
         }
       };